requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.0
botocore==1.34.0
requests==2.31.0
ultralytics>=8.0.196
opencv-python==4.8.1.78
scipy==1.11.4
pandas==2.1.4
numpy==1.24.3